DIVIDE - Passed Div Over. For real

+++ Calculator/calculator.c
@@ -13,8 +13,8 @@ int multiply(const int a, const int b) {
     return a * b;
 }
 
-int divide(const int a, const int b) {
-    if (b == 0) return 0;
+int divide(const float a, const float b) {
+    if (b == 0.0) return 0;
 
     return a / b;
 }
